chore: scaffold the project for the Terraform CDK
Scaffold the project for use with Terraform CDK, using TypeScript as the
programming language and the local repository as cache, by running
`cdktf init --language=TypeScript --local`

The .gitignore file was also updated to ignore common auto-generated Node.JS,
Terraform, and CDKTF files.

Finally, codespell was also configured to ignore the package-lock.json
file, as it kept identifying the inflight JavaScript package as a
spelling mistake.