ci: generate HCL2 file from CDK output
Add a Bash script used to convert the CDK-generated Terraform JSON file
to an HCL2 template. This template can then be consumed by tools like
Checkov.

To convert the Terraform template from JSON to HCL2, we need to follow
the steps listed below:

1. Run `cdktf synth` to generate a Terraform template in JSON
2. Use `jq` to remove null elements, or other elements that could
   interfere with the translation process
3. Use json2hcl to convert the modified JSON template to an HCL1
   compliant template
4. Install version 0.12 of the Terraform CLI to convert the HCL1
   template to an HCL2 template using the 0.12upgrade flag

Once these steps complete, we can pass the new Terraform file to
Checkov, to ensure that the generated cloud resources follow best
practices.

To enable this script for use in the CI pipeline, the $GITHUB_WORKSPACE
environment variable was hardcoded in the paths. This can be revisted in
the future to ensure that this script can be executed locally as well.

For more details, see https://bit.ly/3TciRGV for details on converting
the cdktf JSON templates to HCL2, and https://bit.ly/3zA7RMy for details
on how to use jq to remove comments from the generated JSON object.