ci: run Checkov using the Bridgecrew action
Run Checkov tests using the GitHub Action provided by Bridgecrew. The
action is also configured to publish test results to the Bridgecrew
Cloud project associated to this repository.

Because there are warnings generated by Checkov, to allow the changes to
be merged into the main branch, the scan was configured to soft-fail,
generating warnings but passing the build. Once current issues have been
resolved, the soft-fail parameter can be set to `false`, failing the
build if new issues are introduced in the project.